rename: storeService -> getMembers
