--- src/api/index.ts
@@ -1,2 +1,0 @@
-export type { Grade, Submission } from "./services/common/types";
-export { getMembers } from "./services/store/storeService";
